[REFACT] - Login screen refactoring
