+++ app/src/main/java/project/com/inflix_android/api/dataclass/LoginRequest.kt
@@ -1,0 +1,9 @@
+package project.com.inflix_android.api.dataclass
+
+import com.google.gson.annotations.SerializedName
+import java.io.Serializable
+
+data class LoginRequest(
+    @SerializedName("email") val email : String,
+    @SerializedName("password") val password : String
+) : Serializable
