+++ app/src/main/java/project/com/inflix_android/api/dataclass/LoginResponse.kt
@@ -1,0 +1,9 @@
+package project.com.inflix_android.api.dataclass
+
+import com.google.gson.annotations.SerializedName
+import java.io.Serializable
+
+data class LoginResponse(
+    @SerializedName("success") val success : Boolean,
+    @SerializedName("user") val userResponse : UserResponse
+) : Serializable
