+++ app/src/main/java/project/com/inflix_android/login/presenter/LoginPresenterInterface.kt
@@ -1,4 +1,4 @@
-package project.com.inflix_android.presenter
+package project.com.inflix_android.login.presenter
 
 interface LoginPresenterInterface {
     fun onLogin(email : String, password : String)
