+++ app/src/main/java/project/com/inflix_android/login/view/LoginViewInterface.kt
@@ -1,4 +1,4 @@
-package project.com.inflix_android.view
+package project.com.inflix_android.login.view
 
 interface LoginViewInterface {
     fun onLoginSuccess (message: Int)
