+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frame"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.activity.MainActivity"
+    tools:context=".login.ui.activity.MainActivity"
     tools:layout="@layout/fragment_login">
 
     <fragment
