+++ app/src/main/res/layout/fragment_login.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragment.LoginFragment"
+    tools:context=".login.ui.fragment.LoginFragment"
     android:background="@color/gray"
     app:defaultNavHost="true"
     app:navGraph="@navigation/nav_graph"
